+++ views/site/repository/list.tpl
@@ -3,43 +3,43 @@
   <div class="panel-heading" style="padding-bottom:0px">
     <ul class="breadcrumb">
       <li><a href="/admin">Admin</a></li>
-      <li class="active">Sources</li>
+      <li class="active">Repositories</li>
     </ul>
   </div>
   <div class="panel-body">
     <section>
       <h4>Options</h4>
       <div class="btn-group">
-        <a href="/admin/sources/add" class="btn btn-primary">Add source</a>
+        <a href="/admin/repositories/add" class="btn btn-primary">Add repository</a>
       </div>
     </section>
     <br />
     <section>
-      <h4>Sources</h4>
+      <h4>Repositories</h4>
     {{ if .anything }}
     <div class="table-responsive">
       <table class="table table-striped table-hover ">
         <thead>
           <tr>
             <th>Name</th>
-            <th>URL</th>
+            <th>Base URL</th>
             <th>Description</th>
             <th>Modify</th>
             <th>Delete</th>
           </tr>
         </thead>
         <tbody>
-          {{range $index, $element := .sources}}
+          {{range $index, $element := .repositories}}
           {{$safe := $element.Name|html}}
           <tr>
             <td>{{ $element.Name |html }}</td>
             <td><a href="{{ $element.URL }}" target="_blank">{{ $element.URL |html }}</a></td>
             <td>{{ $element.Description |html }}</td>
             <td>
-              <a href="/admin/sources/modify/{{$element.ID}}">Modify</a>
+              <a href="/admin/repositories/modify/{{$element.ID}}">Modify</a>
             </td>
             <td>
-              <a class="deleteSourceButton" data-id="{{$element.ID}}" data-name="{{$safe}}" href="/">Delete</a>
+              <a class="deleteRepositoryButton" data-id="{{$element.ID}}" data-name="{{$safe}}" href="/">Delete</a>
             </td>
           </tr>
           {{end }}
@@ -47,8 +47,8 @@
       </table>
     </div>
     {{ else }}
-    There are no sources... :(
+    There are no repositories... :(
     {{ end }}
   </div>
 </div>
-<script src="/static/js/admin/source/list.js"></script>
+<script src="/static/js/admin/repository/list.js"></script>
